fix(gui): move stylesheet cache to theme_helpers for CI
