+++ gui_qt/theme.py
@@ -8,17 +8,17 @@ from PySide6.QtWidgets import QApplication
 
 from .font_helpers import GuiFontFamilies, build_font_stylesheet, load_gui_fonts
 from .theme_helpers import (
+    clear_stylesheet_cache,
+    load_theme_stylesheet,
     resolve_effective_theme,
-    theme_stylesheet_filename,
 )
 
-_STYLESHEET_CACHE: dict[tuple[str, str], str] = {}
 _FONT_CACHE: dict[str, GuiFontFamilies] = {}
 
 
 def clear_theme_caches() -> None:
     """Reset module-level caches; primarily intended for test isolation."""
-    _STYLESHEET_CACHE.clear()
+    clear_stylesheet_cache()
     _FONT_CACHE.clear()
 
 
@@ -29,24 +29,6 @@ def system_prefers_dark(app: QApplication) -> bool | None:
     if scheme == Qt.ColorScheme.Light:
         return False
     return None
-
-
-def load_theme_stylesheet(resources_dir: Path, effective_theme: str) -> str:
-    cache_key = (str(resources_dir.resolve()), effective_theme)
-    cached = _STYLESHEET_CACHE.get(cache_key)
-    if cached is not None:
-        return cached
-
-    filename = theme_stylesheet_filename(effective_theme)
-    path = resources_dir / filename
-    if not path.exists() and effective_theme == "dark":
-        path = resources_dir / "app.qss"
-    if not path.exists():
-        stylesheet = ""
-    else:
-        stylesheet = path.read_text(encoding="utf-8")
-    _STYLESHEET_CACHE[cache_key] = stylesheet
-    return stylesheet
 
 
 def cached_gui_fonts(resources_dir: Path) -> GuiFontFamilies:
